chore(workers): isolate casino betting-action display helper (#2126 r5)
- Extract bettingActionKeys + cuiBettingActionName from the shared (core)
  cui_card_helper.go into a new casino-tagged cui_betting_action.go (they
  reference casino domain.PokerAction* and only casino presenters use them).
- Tag poker_presenter_common.go and poker_input_mixins.go casino.
