+++ internal/adapter/presenter/poker_presenter_common.go
@@ -1,3 +1,5 @@
+//go:build !js || !wasm || casino
+
 package presenter
 
 import (
